requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.40.0
google-generativeai==0.8.3
groq==0.13.1
twilio==9.3.2
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
streamlit==1.39.0
apscheduler==3.10.4
python-dotenv==1.0.1
pandas==2.2.3
python-multipart==0.0.12
pytest==8.3.4
httpx==0.28.1
